% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.get_privacy.R
\name{account.get_privacy}
\alias{account.get_privacy}
\title{Return the authenticating user's privacy settings.}
\usage{
account.get_privacy(roauth)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}
}
\value{
A list of: 
 \item{comment}{Who can comment this account's status. 0 : all users; 1 : followings; 2 : credible users.}
 \item{geo}{Whether allow saving and displaying geo info in status. 0 : not allow; 1 : allow.}
 \item{message}{Who can send direct message to this account. 0 : all users; 1 : followings; 2 : credible users.}
 \item{realname}{Whether allow other user find me by real name though the search engine. 0 : not allow; 1 : allow.}
 \item{badge}{Medal status. 0 : private. 1 : public.}
 \item{mobile}{Whether allow other user find me by mobile phone number though the search engine. 0 : not allow; 1 : allow.}
 \item{webim}{Whether enable webim. 0 : no; 1 : yes.}
}
\description{
Return the authenticating user's privacy settings.
}
\examples{
\dontrun{

account.get_privacy(roauth)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/account/get_privacy}
}
\keyword{Accounts}

