% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.getComments.R
\name{analysis.getComments}
\alias{analysis.getComments}
\title{Return the data.frame of comments of a original weibo.}
\usage{
analysis.getComments(roauth, mid, count = -1, startcount = 1)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{mid}{Weibo ID.}

\item{count}{The total count of reposts.}

\item{startcount}{The start number of reposts.}
}
\value{
A data.frame.
}
\description{
Return the data.frame of comments of a original weibo.
}
\examples{
\dontrun{

a1 <- analysis.getComments(roauth, mid = "3575234466298494")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{Analysis}

