% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.destroy.R
\name{comments.destroy}
\alias{comments.destroy}
\title{Delete a comment of authenticating user's.}
\usage{
comments.destroy(roauth, cid, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{cid}{The comment ID to be deleted.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of: 
 \item{created_at}{Created time}
 \item{id}{Weibo ID}
 \item{text}{Weibo content}
 \item{source}{Weibo source}
 \item{mid}{Weibo MID}
 \item{user}{User profile that posted the weibo}
 \item{status}{The weibo that is commented}
 \item{reply_comment}{Replied comment information}
}
\description{
Delete a comment of authenticating user's.
}
\examples{
\dontrun{

com2 <- comments.create(roauth, id = "3543748358960699", 
  comment = "huhu", comment_ori = 0)
comments.destroy(roauth, cid = as.character(com2$id))
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/comments/destroy}
}
\keyword{Comments}

