% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.reply.R
\name{comments.reply}
\alias{comments.reply}
\title{Reply a comment.}
\usage{
comments.reply(roauth, cid, id, comment, without_mention = 0,
  comment_ori = 0, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{cid}{The ID of the comment to be replied.}

\item{id}{The ID of the weibo to be commented.}

\item{comment}{The comment content.It must be encoded by URLEncode within 140 Chinese characters.}

\item{without_mention}{Whether add a mention to the comment content automatically. 0: yes; 1: no; Default is 0.}

\item{comment_ori}{Whether comment the original weibo when comment a reposted weibo. 0: no; 1: yes; Default is 0.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of: 
 \item{created_at}{Created time}
 \item{id}{Weibo ID}
 \item{text}{Weibo content}
 \item{source}{Weibo source}
 \item{mid}{Weibo MID}
 \item{user}{User profile that posted the weibo}
 \item{status}{The weibo that is commented}
 \item{reply_comment}{Replied comment information}
}
\description{
Reply a comment.
}
\examples{
\dontrun{

comments.reply(roauth, cid = "3543759101243075", id = "3543748358960699", comment = "haha")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/comments/reply}
}
\keyword{Comments}

