% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.timeline.R
\name{comments.timeline}
\alias{comments.timeline}
\title{Return the latest comments of the authenticating user including the sending and receiving.}
\usage{
comments.timeline(roauth, count = 20, page = 1, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{count}{The returned count for every page.The default value is 20.}

\item{page}{The returned page.The default value is 1.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of weibos, each weibo contains: 
 \item{idstr}{Weibo ID of string}
 \item{created_at}{Created time}
 \item{id}{Weibo ID}
 \item{text}{Weibo content}
 \item{source}{Weibo source}
 \item{mid}{Weibo MID}
 \item{user}{User profile that posted the weibo}
 \item{status}{The weibo that is commented}
 \item{reply_comment}{Replied comment information}
}
\description{
Return the latest comments of the authenticating user including the sending and receiving.
}
\examples{
\dontrun{

comments.timeline(roauth, count = 5)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/comments/timeline}
}
\keyword{Comments}

