% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.code_to_location.R
\name{common.code_to_location}
\alias{common.code_to_location}
\title{Return address names by address codes.}
\usage{
common.code_to_location(roauth, codes)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{codes}{Address codes to be queried.These are separated by comma.}
}
\value{
A list of address name.
}
\description{
Return address names by address codes.
}
\examples{
\dontrun{

common.code_to_location(roauth, codes = 100)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/common/code_to_location}
}
\keyword{Accounts}

