% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.get_city.R
\name{common.get_city}
\alias{common.get_city}
\title{Return cities list.}
\usage{
common.get_city(roauth, province, capital, language)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{province}{Province code.}

\item{capital}{The first letter of the provinces, a-z. Return all of the provinces 
if the parameter is not provided.Default is not provided.}

\item{language}{The returned language version. zh-cn: Simplified Chinese; 
zh-tw: Traditional Chinese; english: English; Default is zh-cn.}
}
\value{
A list of cities.
}
\description{
Return cities list.
}
\examples{
\dontrun{

common.get_city(roauth, province = "001042")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/common/get_city}
}
\keyword{Accounts}

