% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOAuth.R
\name{createOAuth}
\alias{createOAuth}
\title{create an OAuth object}
\usage{
createOAuth(app_name, access_name, authorize = TRUE, login = FALSE,
  username = "", password = "")
}
\arguments{
\item{app_name}{name of the application.}

\item{access_name}{a string of your access name.}

\item{authorize}{whether to authorize the oauth to use sina API.}

\item{login}{whether to login to impersonate the login.}

\item{username}{user name of the account to impersonate the login.}

\item{password}{password of the account to impersonate the login.}
}
\value{
An reference object of \code{\link{weibo2.0}}.
}
\description{
Create an authorized OAuth object
}
\note{
There is only one OAuth object needed.
}
\examples{
\dontrun{

roauth <- createOAuth("sinademo", "user1")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/OAuth/en}
}
\seealso{
\code{\link{registerApp}}
}
\keyword{authorization}

