% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships.friends.R
\name{friendships.friends}
\alias{friendships.friends}
\title{Return the friends list of a user.}
\usage{
friendships.friends(roauth, uid, screen_name, count = 20, cursor = 0, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{uid}{User ID to be queried.}

\item{screen_name}{User nickname to be queried.}

\item{count}{The returned count for every page.The default value is 20. The max value is 200.}

\item{cursor}{The cursor of the returned data.The returned next_cursor is for the next page.The returned previous_cursor is for the previous page. Default is 0.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of user, which contains: 
 \item{id}{User ID}
 \item{screen_name}{User nickname}
 \item{name}{Friendly displayed name}
 \item{province}{Province ID user located}
 \item{city}{City ID user located}
 \item{location}{Address user located}
 \item{description}{User description}
 \item{url}{Url of the user's blog}
 \item{profile_image_url}{Profile image}
 \item{domain}{The user's personalized weibo url}
 \item{gender}{Gender, m : male; f : female; n : unknown}
 \item{followers_count}{Followers count}
 \item{friends_count}{Friends count}
 \item{statuses_count}{Weibo count}
 \item{favourites_count}{Favorites count}
 \item{created_at}{Created time}
 \item{following}{Whether the current user is following the user}
 \item{allow_all_act_msg}{Whether all the people can send direct message to the user}
 \item{geo_enabled}{Whether can have the geography information}
 \item{verified}{Whether the user is verified by his real identity, marked with "V"}
 \item{allow_all_comment}{Whether all the people can comment the user's weibo}
 \item{avatar_large}{Profile large image}
 \item{verified_reason}{verification reason}
 \item{follow_me}{Whether the user is following the current user}
 \item{online_status}{Whether the user is online, 0 : offline, 1 : online}
 \item{bi_followers_count}{The count of the users that are following the user and are being followed by the user}
 \item{status}{The latest weibo of the user}
}
\description{
Return the friends list of a user.
}
\note{
You must provide a parameter either uid or screen_name.
}
\examples{
\dontrun{

friendships.friends(roauth, uid = "1318558807")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/friendships/friends}
}
\keyword{Friendships}

