\name{Rweibo-package}
\alias{Rweibo-package}
\alias{Rweibo-package}
\title{Provides an interface to the Weibo open platform}
\description{This package provide an interface to sina weibo using OAuth}
\details{\tabular{ll}{ 
Package: \tab Rweibo\cr 
Type: \tab Package\cr 
Version: \tab 0.0-1\cr
Date: \tab 2011-04-16\cr 
License: \tab GPL\cr 
LazyLoad: \tab yes\cr 
} 

Weibo open platform is an information subscription, sharing and communication platform based on Sina Weibo. Weibo open platform provides communication channel with huge information, friendship.

On the open platform website, developers or websites can create own creative application to access the information in Weibo system through our Open API.}
\alias{Rweibo-package}
\alias{Rweibo}
\docType{package}
\author{Jian li

Maintainer: Jian Li <lijian.pku@gmail.com>}
\references{\url{http://open.weibo.com/}}
\keyword{package}

