% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.destroy.R
\name{statuses.destroy}
\alias{statuses.destroy}
\title{Delete a weibo by the ID.}
\usage{
statuses.destroy(roauth, id, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{id}{The weibo ID to be deleted.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of weibos, each weibo contains: 
 \item{idstr}{Weibo ID of string}
 \item{created_at}{Created time}
 \item{id}{Weibo ID}
 \item{text}{Weibo content}
 \item{source}{Weibo source}
 \item{favorited}{Whether it is favorited}
 \item{truncated}{Whether it is truncated}
 \item{in_reply_to_status_id}{The reply's ID}
 \item{in_reply_to_user_id}{Replyer ID}
 \item{in_reply_to_screen_name}{Replyer nickname}
 \item{mid}{Weibo MID}
 \item{bmiddle_pic}{Medium picture}
 \item{original_pic}{Original picture}
 \item{thumbnail_pic}{Thumbnail picture}
 \item{reposts_count}{Repost count}
 \item{comments_count}{Comment count}
 \item{annotations}{Weibo annotation information}
 \item{geo}{Geography information}
 \item{user}{User profile that posted the weibo}
}
\description{
Delete a weibo by the ID.
}
\note{
You can only delete the weibo created by yourself.
}
\examples{
\dontrun{

rep1 <- statuses.repost(roauth, id = "3543748358960699", status = "hello")
statuses.destroy(roauth, id = rep1$idstr)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/statuses/destroy}
}
\keyword{Weibo}

