% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.mentions.R
\name{statuses.mentions}
\alias{statuses.mentions}
\title{Return the latest N weibos metioned the authenticating user. (That is the weibo contained @username)}
\usage{
statuses.mentions(roauth, count = 20, page = 1, filter_by_author = 0,
  filter_by_source = 0, filter_by_type = 0, ...)
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{count}{The returned count for every page.The default value is 20.}

\item{page}{The returned page.The default value is 1.}

\item{filter_by_author}{Return the weibos by the owner type. 0 : all types; 1 : my friend; 2 : stranger; The default value is 0.}

\item{filter_by_source}{Return the weibos by the source type. 0 : all types; 1 : from weibo; 2 : from weiqun; The default value is 0.}

\item{filter_by_type}{Return the weibos by the type. 0 : all types; 1 : original weibos; The default value is 0.}

\item{...}{Other request parameters for this API.}
}
\value{
A list of weibos, each weibo contains: 
 \item{idstr}{Weibo ID of string}
 \item{created_at}{Created time}
 \item{id}{Weibo ID}
 \item{text}{Weibo content}
 \item{source}{Weibo source}
 \item{favorited}{Whether it is favorited}
 \item{truncated}{Whether it is truncated}
 \item{in_reply_to_status_id}{The reply's ID}
 \item{in_reply_to_user_id}{Replyer ID}
 \item{in_reply_to_screen_name}{Replyer nickname}
 \item{mid}{Weibo MID}
 \item{bmiddle_pic}{Medium picture}
 \item{original_pic}{Original picture}
 \item{thumbnail_pic}{Thumbnail picture}
 \item{reposts_count}{Repost count}
 \item{comments_count}{Comment count}
 \item{annotations}{Weibo annotation information}
 \item{geo}{Geography information}
 \item{user}{User profile that posted the weibo}
}
\description{
Return the latest N weibos metioned the authenticating user. (That is the weibo contained @username)
}
\examples{
\dontrun{

statuses.mentions(roauth, count = 5)
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki/2/statuses/mentions}
}
\keyword{Weibo}

