% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.search.content.R
\name{web.search.content}
\alias{web.search.content}
\title{Search content by word.}
\usage{
web.search.content(sword, page = 1, combinewith = NULL, since = NULL,
  sinceID = NULL, sleepmean = 3, sleepsd = 1, roauth = NULL, ...)
}
\arguments{
\item{sword}{The word to be searched.}

\item{page}{The number of the total pages of the search result. The default value is 1. The limited value is 50.}

\item{combinewith}{A data.frame of the previous search result. The default value is NULL.}

\item{since}{If not NULL, restricts weibos to those since the given date. Date should be an object of class \code{\link{POSIXlt}} or string formatted as YYYY-MM-DD}

\item{sinceID}{If not NULL, returns weibos with IDs greater (ie newer) than the specified ID.}

\item{sleepmean}{Mean of the sleeping time before each searching.}

\item{sleepsd}{Standard deviations of the sleeping time before each searching.}

\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{...}{Other request parameters for this API.}
}
\value{
A data.frame of weibos, each weibo contains: 
 \item{MID}{Weibo ID of string}
 \item{Author}{Nick name of the author}
 \item{Weibo}{Weibo content}
 \item{Forward}{Forward content}
 \item{Time_Weibo}{Time of this weibo}
 \item{Time_Search}{Time of this search}
 \item{Count_Forward}{Count of replys}
 \item{Count_Reply}{Count of forwards}
}
\description{
Search content by word.
}
\examples{
\dontrun{
web.search.content("Rweibo")$Weibo
web.search.content("Rweibo", sinceID = "3508307023192146")$Weibo
web.search.content("Rweibo", since = "2012-10-01")$Weibo
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{Search}

