% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.search.user.R
\name{web.search.user}
\alias{web.search.user}
\title{Search user by nick name.}
\usage{
web.search.user(screen_name, roauth = NULL)
}
\arguments{
\item{screen_name}{Nick name of a user.}

\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}
}
\value{
A data.frame of weibos, each weibo contains: 
 \item{uid}{User ID}
 \item{url}{URL of user}
 \item{follow}{Number of followers}
 \item{fan}{Number of funs}
 \item{profile}{Number of weibos}
}
\description{
Search user by nick name.
}
\examples{
\dontrun{
web.search.user("lijian001")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{Search}

