% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibo.api.R
\name{weibo.api}
\alias{weibo.api}
\title{Invoke an API directly by it's URL.}
\usage{
weibo.api(roauth, URL, paramlist = list(), httpmethod = c("GET", "POST"))
}
\arguments{
\item{roauth}{A OAuth object created by \code{\link{createOAuth}}.}

\item{URL}{The url of this API.}

\item{paramlist}{Parameters list.}

\item{httpmethod}{HTTP request method, 'GET' or 'POST'.}
}
\value{
A R object.
}
\description{
Invoke an API directly by it's URL.
}
\examples{
\dontrun{

weibo.api(roauth, URL = "https://api.weibo.com/2/statuses/public_timeline.json", 
    paramlist = list(count = 5), httpmethod = "GET")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{http://open.weibo.com/wiki}
}
\keyword{Weibo}

