\name{weibo.update}
\alias{weibo.update}
\title{Update the Rweibo package from R-forge automatically.}
\usage{
weibo.update()
}
\value{
Invisible TRUE or FALSE.
}
\description{
Update the Rweibo package from R-forge automatically.
}
\examples{
\dontrun{

weibo.update()
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{Weibo}

