\name{weibo.checkUpdate}
\alias{weibo.checkUpdate}
\title{Check whether there is a new version of Rweibo.}
\usage{
weibo.checkUpdate()
}
\value{
Invisible TRUE or FALSE.
}
\description{
Check whether there is a new version of Rweibo.
}
\examples{
\dontrun{

weibo.checkUpdate()
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{Weibo}

